+++ Servicios/Modelos/InfoTabla.cs
@@ -1,0 +1,7 @@
+namespace ABD_Project.Servicios.Modelos;
+
+/// <summary>
+/// Información mínima de una tabla para el diagrama.
+/// </summary>
+/// <param name="Nombre">Nombre lógico de la tabla.</param>
+public record InfoTabla(string Nombre);
