+++ Servicios/Renderizadores/ConstructorDiagramaChen.cs
@@ -1,8 +1,11 @@
-﻿using System;
+using ABD_Project.Servicios.Modelos;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
+
+namespace ABD_Project.Servicios.Renderizadores;
 
 /// <summary>
 /// Construye un diagrama ER en notación Chen usando sintaxis de Mermaid (flowchart),
